requirements.txt
matplotlib
opencv-python
Pillow
torch
transformers
